% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/euler_maruyama.R
\name{euler_maruyama}
\alias{euler_maruyama}
\title{Implements the Euler-Maruyama method to solve the SDE
dX_t = drift(t,X_t) dt + dispersion(t,X_t) dW_t}
\usage{
euler_maruyama(t_max, h, drift, dispersion, X0)
}
\arguments{
\item{t_max}{the time horizon of the simulation.
the path of the process will be estimated on the interval
[0,floor(t_max/h)*h]}

\item{h}{the step size taken}

\item{drift}{the drift term
(should accept arguments of the form (t,X_t))}

\item{dispersion}{the dispersion term (sigma)
(should accept arguments of the form (t,X_t))}

\item{X0}{initial conditions}
}
\description{
Implements the Euler-Maruyama method to solve the SDE
dX_t = drift(t,X_t) dt + dispersion(t,X_t) dW_t
}

