% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/euler_maruyama.R
\name{testEM}
\alias{testEM}
\title{A script for testing the EM-mathod}
\usage{
testEM(N_paths = 20, t_max = 1, h = 0.01)
}
\arguments{
\item{N_paths}{the number of paths to be sampled}

\item{t_max}{the time-horizon of simulations}

\item{h}{the step size.}
}
\description{
Simulates a geometric brownian motion, and plots the results.
The number of paths sampled is \code{N_paths}; the number of steps taken
per path is \code{floor(t_max / h)}.
}

